ISPN-4415 Provide correct Managed executor facade and fix XSLT
* Correct executor thread factories needed to be provided according
to the executor being plugged. Scheduled ones need a different facade.
* Fix an XSLT issue so that custom Hot Rod endpoints are added
correctly.
