+++ server/integration/testsuite/src/test/resources/config/xslt/config-modifier.xsl
@@ -1,6 +1,6 @@
 <xsl:stylesheet version="1.0"
                 xmlns:xsl="http://www.w3.org/1999/XSL/Transform"
-                xmlns:p="urn:jboss:domain:1.4"
+                xmlns:p="urn:jboss:domain:2.1"
                 xmlns:logging="urn:jboss:domain:logging:1.2">
     <xsl:output method="xml" indent="yes" omit-xml-declaration="yes"/>
